+++ CHANGES.md
@@ -1,3 +1,2 @@
-- Updated for patch 10.2
-- Updated for MDT 4.2.0.*
-- Fixed trying to scroll to pull when no pull data is available
+- Added command to disable map animations
+- Removed height setting command